A manager component added to create a group easily and configuration input changed
The manager component creates a prober which object which compries of
multiple pwm_lines object
The configuration input settings are changed to be more user friendly.
Now it takes in microseconds units for time_period, pulse width and
dead time. The pulse width is the actual width of the pulse generateed